support GUI Display Data on 0d variables
A 0 dimention numpy array is weird.  It's neither a scalar, nor a 1 dimensional array with a single element. "Display Data" silently refused to work for such values.

This change converts the 0d array into a 1d array with a single element immediately before trying to display it.  Result: Display Data now works on 0d arrays, showing the single value as expected.