when gui loads a file, guess at default variable
If we had a file open in that "slot" (A or B) already, and the selected variable name is also present in the new file, pick that.
If that doesn't work, check to see if the selected variable name in the other file (B or A) is present, and if so pick that.
Otherwise stick with the existing default: the first item.

This should speed things up a little bit when bouncing between files, as the variable the user is most likely interested will already be selected.